requirements.txt
flask>=2.0.0
ollama
python-dotenv
marshmallow
requests
logging